// frontend/src/interfaces/vehicleModel.ts
export interface vehicleModel {
    vehicleId: number
    brandAndModel: string
    manufacturingYear: number
    color: string
    engineCapacity: number
    price: number
    millage: number
    plateNumber: string
    isActive: boolean
    auctionPrice: number
    additionalInformation: string
    startTime: string
    endTime: string
    image: string
    sellerId: string
    bids: any
  }
  